Feature/notification domain

+++ src/main/java/com/example/doktoribackend/user/policy/ReadingPurpose.java
@@ -24,7 +24,7 @@ public class ReadingPurpose extends BaseTimeEntity {
     private String name;
 
     @Column(name= "priority", nullable = false, columnDefinition = "TINYINT")
-    private Byte priority;
+    private Integer priority;
 
     @Column(name = "deleted_at")
     private LocalDateTime deletedAt;
